+++ examples/async_menagerie.rs
@@ -159,8 +159,8 @@ impl Task {
                 .enq_async().unwrap()
         };
 
-        if is_write { future_data.set_unmap_wait_list(unmap_wait_list.unwrap()); }
-        let unmap_event_target = future_data.create_unmap_completion_event().unwrap().clone();
+        if is_write { future_data.set_unmap_wait_events(unmap_wait_list.unwrap()); }
+        let unmap_event_target = future_data.create_unmap_event().unwrap().clone();
         self.cmd_graph.set_cmd_event(cmd_idx, unmap_event_target.into()).unwrap();
 
         future_data