+++ src/tests/async.rs
@@ -138,7 +138,7 @@ pub fn write_init(
 
     let mut future_guard = rw_vec.clone().write();
     // future_guard.set_wait_event(wait_marker.as_ref().unwrap().clone());
-    future_guard.set_wait_list(wait_list);
+    future_guard.set_lock_wait_events(wait_list);
     let release_event = future_guard.create_release_event(write_init_release_queue_0)
         .unwrap().clone();